# openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/customers": {
      "get": {
        "operationId": "CustomersController_getAllCustomers",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appId",
            "required": false,
            "in": "query",
            "description": "App ID filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          }
        },
        "summary": "Get all customers with pagination",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/customers/{id}": {
      "get": {
        "operationId": "CustomersController_getCustomerById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "summary": "Get customer by ID",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/customers/{id}/subscriptions": {
      "get": {
        "operationId": "CustomersController_getCustomerSubscriptions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer subscriptions"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "summary": "Get customer subscriptions",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/api/privacy-policies": {
      "get": {
        "operationId": "PublicPrivacyPoliciesController_renderPoliciesList",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicPrivacyPolicies"
        ]
      },
      "post": {
        "operationId": "PrivacyPoliciesController_createPolicy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrivacyPolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Privacy policy created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Privacy policy with this ID already exists"
          }
        },
        "summary": "Create a new privacy policy",
        "tags": [
          "privacy-policies"
        ]
      }
    },
...
    "/api/privacy-policies/{id}": {
      "get": {
        "operationId": "PrivacyPoliciesController_getPolicyById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Get privacy policy by ID",
        "tags": [
          "privacy-policies"
        ]
      },
      "put": {
        "operationId": "PrivacyPoliciesController_updatePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrivacyPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Privacy policy updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Update an existing privacy policy",
        "tags": [
          "privacy-policies"
        ]
      },
      "delete": {
        "operationId": "PrivacyPoliciesController_deletePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy deleted successfully"
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Delete a privacy policy",
        "tags": [
          "privacy-policies"
        ]
      }
    },
...
    "/api/privacy-policy/{id}": {
      "get": {
        "operationId": "PublicPrivacyPoliciesController_renderPolicyPage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicPrivacyPolicies"
        ]
      }
    },
    "/api/privacy-policy/app/{appId}": {
      "get": {
        "operationId": "PublicPrivacyPoliciesController_renderPolicyPageByAppId",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicPrivacyPolicies"
        ]
      }
    },
    "/api/privacy-policy/app/name/{appName}": {
      "get": {
        "operationId": "PublicPrivacyPoliciesController_renderPolicyPageByAppName",
        "parameters": [
          {
            "name": "appName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicPrivacyPolicies"
        ]
      }
    },
...
  },
  "info": {
    "title": "Doctopus API",
    "description": "Modern NestJS API with comprehensive features",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://tree2u.kobotogether.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "PrivacyPolicyDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePrivacyPolicyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePrivacyPolicyDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}